12-1

+++ Programacion/Actividades/src/com/dam/programame/p116.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class p116 {
     public static void main(String[] args) {
         Scanner sc=new Scanner(System.in);
-        int n=Integer.parseInt(sc.nextLine());
+        byte n=Integer.parseInt(sc.nextLine());
         sc.close();
-        for(int i=0; i<n; i++){
+        for(byte i=0; i<n; i++){
             System.out.println("Hola mundo.");
         }
     }
